+++ zdc/include/zd/gen/zd4_section.hpp
@@ -1,3 +1,5 @@
+#pragma once
+
 #include <cstdint>
 #include <cstdio>
 #include <list>
