+++ zdc/src/zd/gen/CMakeLists.txt
@@ -1,4 +1,5 @@
 target_sources(zdc PRIVATE text_generator.cpp)
+target_sources(zdc PRIVATE x86_assembler.cpp)
 target_sources(zdc PRIVATE zd4_builtins.cpp)
 target_sources(zdc PRIVATE zd4_generator.cpp)
 target_sources(zdc PRIVATE zd4_section.cpp)
